png_documents_validator/requirements.txt
fastapi
uvicorn
pillow
pytesseract
paddlepaddle
paddleocr
python-multipart
pyzbar
pdf2image
